# docs/api-contracts/user-service-openapi.yaml
openapi: 3.0.3
info:
  title: User Service API
  description: User microservice for Global Artisan Marketplace - auth and profiles
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local
  - url: https://user-service.example.com
    description: Production

paths:
  /api/auth/register:
    post:
      summary: Register a new buyer or seller
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error or email already exists

  /api/auth/login:
    post:
      summary: Login and receive JWT token
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid credentials

  /api/users/{id}:
    get:
      summary: Get user profile (for inter-service integration)
      tags: [User]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '404':
          description: User not found

    put:
      summary: Update own profile
      tags: [User]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated profile
        '403':
          description: Forbidden - can only update own profile

  /api/users/{id}/validate:
    get:
      summary: Validate user exists (internal - for Order/Review service)
      tags: [User]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  valid:
                    type: boolean

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        displayName:
          type: string
          maxLength: 100
        role:
          type: string
          enum: [BUYER, SELLER]
        country:
          type: string

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        userId:
          type: string
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
        expiresIn:
          type: integer

    UserProfileResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
        country:
          type: string

    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
        avatarUrl:
          type: string
        country:
          type: string