# user-service/src/main/java/com/artisan/user/dto/AuthResponse.java
package com.artisan.user.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class AuthResponse {
    private String token;
    private String userId;
    private String email;
    private String displayName;
    private String role;
    private long expiresIn;
}